+++ lib/auction_fun_core/entities/user.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module AuctionFunCore
+  module Entities
+    # User Relations class. This return simple objects with attribute readers
+    # to represent data in your user.
+    class User < ROM::Struct
+      def info
+        attributes.except(:password_digest)
+      end
+    end
+  end
+end