+++ spec/spec_helper.rb
@@ -1,21 +1,27 @@
 # frozen_string_literal: true
 
-ENV['APP_ENV'] = 'test'
+ENV["APP_ENV"] = "test"
 
-if ENV['CI']
-  require 'simplecov'
+if ENV["CI"]
+  require "simplecov"
   SimpleCov.start do
-    add_filter '/spec'
-    add_group 'System', 'system'
-    add_group 'Config', 'config'
+    add_filter "/spec"
+    add_group "Commands", "lib/auction_fun_core/commands"
+    add_group "Contracts", "lib/auction_fun_core/contracts"
+    add_group "Entities", "lib/auction_fun_core/entities"
+    add_group "Operations", "lib/auction_fun_core/operations"
+    add_group "Relations", "lib/auction_fun_core/relations"
+    add_group "Repositories", "lib/auction_fun_core/repos"
+    add_group "System", "system"
+    add_group "Config", "config"
   end
 end
 
-require_relative '../config/application'
-require 'pry'
-require 'dotenv'
-require 'rom-factory'
-require 'database_cleaner/sequel'
+require_relative "../config/application"
+require "pry"
+require "dotenv"
+require "rom-factory"
+require "database_cleaner/sequel"
 
 AuctionFunCore::Application.start(:core)
 
@@ -38,4 +44,14 @@ RSpec.configure do |config|
   config.expect_with :rspec do |c|
     c.syntax = :expect
   end
+
+  config.before do
+    DatabaseCleaner.clean
+  end
+
+  config.after(:suite) do
+    AuctionFunCore::Application.stop(:core)
+  end
 end
+
+DatabaseCleaner.strategy = :truncation